refactor: testフォルダ内の保管方法を変更
